+++ src/SafeRegex/Exception/CatastrophicBacktrackingPregException.php
@@ -1,7 +1,7 @@
 <?php
 namespace TRegx\SafeRegex\Exception;
 
-class CatastrophicBacktrackingPregException extends RuntimePregException
+class CatastrophicBacktrackingPregException extends RuntimePregException implements PatternStructureException
 {
     public function __construct(string $message, $pattern, string $methodName, int $errorCode, string $errorName)
     {