+++ src/SafeRegex/Exception/JitStackLimitPregException.php
@@ -1,7 +1,7 @@
 <?php
 namespace TRegx\SafeRegex\Exception;
 
-class JitStackLimitPregException extends RuntimePregException
+class JitStackLimitPregException extends RuntimePregException implements PatternStructureException
 {
     public function __construct(string $message, $pattern, string $methodName, int $errorCode, string $errorName)
     {