+++ src/SafeRegex/Exception/PatternStructureException.php
@@ -1,0 +1,6 @@
+<?php
+namespace TRegx\SafeRegex\Exception;
+
+interface PatternStructureException extends PregException
+{
+}